// frontend/src/lib/k8s/gen/model/IoK8sApiCoreV1Namespace.ts
export interface IoK8sApiCoreV1Namespace {
  /**
   * APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
   * @type {string}
   * @memberof IoK8sApiCoreV1Namespace
   */
  apiVersion?: string;
  /**
   * Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
   * @type {string}
   * @memberof IoK8sApiCoreV1Namespace
   */
  kind?: string;
  /**
   *
   * @type {IoK8sApimachineryPkgApisMetaV1ObjectMeta}
   * @memberof IoK8sApiCoreV1Namespace
   */
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  /**
   *
   * @type {IoK8sApiCoreV1NamespaceSpec}
   * @memberof IoK8sApiCoreV1Namespace
   */
  spec?: IoK8sApiCoreV1NamespaceSpec;
  /**
   *
   * @type {IoK8sApiCoreV1NamespaceStatus}
   * @memberof IoK8sApiCoreV1Namespace
   */
  status?: IoK8sApiCoreV1NamespaceStatus;
}