// frontend/src/lib/k8s/gen/model/IoK8sApimachineryPkgApisMetaV1ObjectMeta.ts
export interface IoK8sApimachineryPkgApisMetaV1ObjectMeta {
// ...
  namespace?: string;
// ...
}